refactor: make code cleaner
