+++ crate/src/pipeline/spatial_ops/mod.rs
@@ -1,4 +1,4 @@
-use crate::pipeline::{Pipeline, PlanarImage};
+use crate::pipeline::Pipeline;
 
 #[macro_use]
 pub(crate) mod common;
@@ -10,16 +10,11 @@ mod sobel_global;
 
 use common::restore_alpha_if_filter_zeroed_it;
 
+use crate::pipeline::spatial_ops::direct_3x3::convolve_3x3;
+use crate::pipeline::spatial_ops::separable::box_blur_simd::box_blur_3x3_simd;
+use crate::pipeline::spatial_ops::separable::convolve_separable_3x3;
 #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
 use direct_3x3::edge_one_simd::edge_one_simd;
-#[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-use separable::box_blur_simd::box_blur_3x3_simd;
-#[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-use separable::line_detection::{
-    detect_horizontal_lines_simd, detect_vertical_lines_simd,
-};
-#[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-use separable::sobel_simd::{sobel_horizontal_simd, sobel_vertical_simd};
 
 #[cfg(not(all(target_arch = "wasm32", target_feature = "simd128")))]
 const NOISE_REDUCTION: [f32; 9] = [0.0, -1.0, 7.0, -1.0, 5.0, 9.0, 0.0, 7.0, 9.0];
@@ -129,15 +124,7 @@ impl Pipeline {
     pub fn detect_horizontal_lines(mut self) -> Self {
         #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
         unsafe {
-            self.flush_pixel_ops();
-            self.ensure_scratch();
-            self.ensure_i16_scratch();
-
-            let scratch = self.scratch.as_mut().unwrap();
-            let i16_scratch = self.i16_scratch.as_mut().unwrap();
-
-            detect_horizontal_lines_simd(&self.image, scratch, i16_scratch);
-            std::mem::swap(&mut self.image, scratch);
+            apply_direct_3x3_simd!(self, [-1, -1, -1, 2, 2, 2, -1, -1, -1]);
         }
 
         #[cfg(not(all(target_arch = "wasm32", target_feature = "simd128")))]
@@ -151,13 +138,7 @@ impl Pipeline {
     pub fn detect_vertical_lines(mut self) -> Self {
         #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
         unsafe {
-            self.flush_pixel_ops();
-            self.ensure_scratch();
-
-            let scratch = self.scratch.as_mut().unwrap();
-
-            detect_vertical_lines_simd(&self.image, scratch);
-            std::mem::swap(&mut self.image, scratch);
+            apply_direct_3x3_simd!(self, [-1, 2, -1, -1, 2, -1, -1, 2, -1]);
         }
 
         #[cfg(not(all(target_arch = "wasm32", target_feature = "simd128")))]
@@ -247,15 +228,7 @@ impl Pipeline {
     pub fn sobel_horizontal(mut self) -> Self {
         #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
         unsafe {
-            self.flush_pixel_ops();
-            self.ensure_scratch();
-            self.ensure_i16_scratch();
-
-            let scratch = self.scratch.as_mut().unwrap();
-            let i16_scratch = self.i16_scratch.as_mut().unwrap();
-
-            sobel_horizontal_simd(&self.image, scratch, i16_scratch);
-            std::mem::swap(&mut self.image, scratch);
+            apply_direct_3x3_simd!(self, [-1, -2, -1, 0, 0, 0, 1, 2, 1]);
         }
 
         #[cfg(not(all(target_arch = "wasm32", target_feature = "simd128")))]
@@ -283,15 +256,7 @@ impl Pipeline {
     pub fn sobel_vertical(mut self) -> Self {
         #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
         unsafe {
-            self.flush_pixel_ops();
-            self.ensure_scratch();
-            self.ensure_i16_scratch();
-
-            let scratch = self.scratch.as_mut().unwrap();
-            let i16_scratch = self.i16_scratch.as_mut().unwrap();
-
-            sobel_vertical_simd(&self.image, scratch, i16_scratch);
-            std::mem::swap(&mut self.image, scratch);
+            apply_direct_3x3_simd!(self, [-1, 0, 1, -2, 0, 2, -1, 0, 1]);
         }
 
         #[cfg(not(all(target_arch = "wasm32", target_feature = "simd128")))]
@@ -335,135 +300,5 @@ impl Pipeline {
 
         convolve_separable_3x3(&self.image, scratch, f32_scratch, horizontal, vertical);
         std::mem::swap(&mut self.image, scratch);
-    }
-}
-
-pub fn convolve_3x3(src: &PlanarImage, dst: &mut PlanarImage, kernel: [f32; 9]) {
-    let width = src.width() as usize;
-    let height = src.height() as usize;
-
-    // If we want to match the result from the photon library, then we skip the borders
-    // So we must fill the lanes with 0 beforehand so that we do not have accidental stale data in there
-    dst.r.fill(0);
-    dst.g.fill(0);
-    dst.b.fill(0);
-    dst.a.fill(0);
-
-    convolve_3x3_channel(&src.r, &mut dst.r, width, height, kernel);
-    convolve_3x3_channel(&src.g, &mut dst.g, width, height, kernel);
-    convolve_3x3_channel(&src.b, &mut dst.b, width, height, kernel);
-    convolve_3x3_channel(&src.a, &mut dst.a, width, height, kernel);
-    restore_alpha_if_filter_zeroed_it(src, dst, width, height);
-}
-
-fn convolve_3x3_channel(
-    src: &[u8],
-    dst: &mut [u8],
-    width: usize,
-    height: usize,
-    kernel: [f32; 9],
-) {
-    if width < 3 || height < 3 {
-        return;
-    }
-
-    let sum: f32 = kernel.iter().sum();
-    let divisor = if sum == 0.0 { 1.0 } else { sum };
-
-    for y in 1..height - 1 {
-        let top = (y - 1) * width;
-        let mid = y * width;
-        let bot = (y + 1) * width;
-
-        for x in 1..width - 1 {
-            let value = src[top + x - 1] as f32 * kernel[0]
-                + src[top + x] as f32 * kernel[1]
-                + src[top + x + 1] as f32 * kernel[2]
-                + src[mid + x - 1] as f32 * kernel[3]
-                + src[mid + x] as f32 * kernel[4]
-                + src[mid + x + 1] as f32 * kernel[5]
-                + src[bot + x - 1] as f32 * kernel[6]
-                + src[bot + x] as f32 * kernel[7]
-                + src[bot + x + 1] as f32 * kernel[8];
-
-            dst[mid + x] = (value / divisor).clamp(0.0, 255.0) as u8;
-        }
-    }
-}
-
-pub fn convolve_separable_3x3(
-    src: &PlanarImage,
-    dst: &mut PlanarImage,
-    scratch: &mut [f32],
-    horizontal: [f32; 3],
-    vertical: [f32; 3],
-) {
-    let width = src.width() as usize;
-    let height = src.height() as usize;
-
-    dst.r.fill(0);
-    dst.g.fill(0);
-    dst.b.fill(0);
-    dst.a.fill(0);
-
-    convolve_separable_3x3_channel(
-        &src.r, &mut dst.r, scratch, width, height, horizontal, vertical,
-    );
-    convolve_separable_3x3_channel(
-        &src.g, &mut dst.g, scratch, width, height, horizontal, vertical,
-    );
-    convolve_separable_3x3_channel(
-        &src.b, &mut dst.b, scratch, width, height, horizontal, vertical,
-    );
-    convolve_separable_3x3_channel(
-        &src.a, &mut dst.a, scratch, width, height, horizontal, vertical,
-    );
-    restore_alpha_if_filter_zeroed_it(src, dst, width, height);
-}
-
-fn convolve_separable_3x3_channel(
-    src: &[u8],
-    dst: &mut [u8],
-    scratch: &mut [f32],
-    width: usize,
-    height: usize,
-    horizontal: [f32; 3],
-    vertical: [f32; 3],
-) {
-    if width < 3 || height < 3 {
-        return;
-    }
-
-    let divisor = {
-        let sum = horizontal.iter().sum::<f32>() * vertical.iter().sum::<f32>();
-        if sum == 0.0 {
-            1.0
-        } else {
-            sum
-        }
-    };
-
-    for y in 0..height {
-        let row = y * width;
-
-        for x in 1..width - 1 {
-            scratch[row + x] = src[row + x - 1] as f32 * horizontal[0]
-                + src[row + x] as f32 * horizontal[1]
-                + src[row + x + 1] as f32 * horizontal[2];
-        }
-    }
-
-    for y in 1..height - 1 {
-        let top = (y - 1) * width;
-        let mid = y * width;
-        let bot = (y + 1) * width;
-
-        for x in 1..width - 1 {
-            let value = scratch[top + x] * vertical[0]
-                + scratch[mid + x] * vertical[1]
-                + scratch[bot + x] * vertical[2];
-
-            dst[mid + x] = (value / divisor).clamp(0.0, 255.0) as u8;
-        }
     }
 }
